requirements.txt
# python==3.8.10
albumentations==1.3.0
Flask==3.0.3
Flask-Cors==5.0.0
numpy==1.21.6
opencv-python==4.11.0.86
python-dotenv==1.0.1
reportlab==4.4.1
SimpleITK==2.2.1
torch==1.13.1
torchvision==0.14.1